Making prev and next buttons properly clickable
The prev and next buttons were not entirely clickable. You had to click
the link text in the middle but could not click anywhere within the
button.

The reason is that the prev and next links were implemented as LI tags
and not just A tags.

See: https://bulma.io/documentation/components/pagination/